adds prepareRecordExport + changes to lastRecordExportCredentials
