+++ app.py
@@ -31,7 +31,6 @@ from fast_api.routes.lookup_lists import router as lookup_lists_router
 from fast_api.routes.heuristic import router as heuristic_router
 from fast_api.routes.data_browser import router as data_browser_router
 from fast_api.routes.labeling import router as labeling_router
-from fast_api.routes.transfer import router as transfer_router
 from middleware.database_session import DatabaseSessionHandler
 from starlette.applications import Starlette
 from starlette.graphql import GraphQLApp
@@ -56,7 +55,6 @@ from route_prefix import (
     PREFIX_HEURISTIC,
     PREFIX_DATA_BROWSER,
     PREFIX_LABELING,
-    PREFIX_TRANSFER,
 )
 from util import security, clean_up
 
@@ -99,7 +97,6 @@ fastapi_app.include_router(
     data_browser_router, prefix=PREFIX_DATA_BROWSER, tags=["data-browser"]
 )
 fastapi_app.include_router(labeling_router, prefix=PREFIX_LABELING, tags=["labeling"])
-fastapi_app.include_router(transfer_router, prefix=PREFIX_TRANSFER, tags=["transfer"])
 
 routes = [
     Route(
