+++ middleware/database_session.py
@@ -17,7 +17,6 @@ from route_prefix import (
     PREFIX_HEURISTIC,
     PREFIX_DATA_BROWSER,
     PREFIX_LABELING,
-    PREFIX_TRANSFER,
 )
 from submodules.model.business_objects import general
 from controller.auth import manager as auth_manager
@@ -35,7 +34,6 @@ PROJECT_ACCESS_PREFIX = [
     PREFIX_HEURISTIC,
     PREFIX_DATA_BROWSER,
     PREFIX_LABELING,
-    PREFIX_TRANSFER,
 ]
