+++ route_prefix.py
@@ -13,4 +13,3 @@ PREFIX_LOOKUP_LISTS = PREFIX + "/lookup-lists"
 PREFIX_HEURISTIC = PREFIX + "/heuristic"
 PREFIX_DATA_BROWSER = PREFIX + "/data-browser"
 PREFIX_LABELING = PREFIX + "/labeling"
-PREFIX_TRANSFER = PREFIX + "/transfer"
